requirements.txt
pandas
numpy
scikit-learn
pyyaml
pytest

